# frontend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "MoodBite API",
    "description": "Gợi ý nhà hàng theo ngữ cảnh: câu tìm kiếm tự do + vị trí + thời điểm.\n\nMọi response bọc trong `data` (thành công) hoặc `error` (lỗi).\nXem GET /api/v1/health để biết nguồn dữ liệu nào đã sẵn sàng.",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/me/addresses": {
      "get": {
        "tags": [
          "me"
        ],
        "summary": "List Addresses",
        "description": "Địa chỉ đã lưu. Địa chỉ MẶC ĐỊNH đứng đầu (nếu có).",
        "operationId": "list_addresses_api_v1_me_addresses_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AddressesResponse"
                }
              }
            }
          },
          "400": {
            "description": "INVALID_REQUEST",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "404": {
            "description": "COLLECTION_NOT_FOUND | ADDRESS_NOT_FOUND",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "503": {
            "description": "DATA_NOT_READY",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "500": {
            "description": "INTERNAL_ERROR",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "401": {
            "description": "UNAUTHORIZED",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "me"
        ],
        "summary": "Create Address",
        "description": "Toạ độ ngoài Hà Nội -> 400 kèm câu giải thích (MoodBite chỉ có dữ liệu Hà Nội).",
        "operationId": "create_address_api_v1_me_addresses_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateAddressRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserAddressResponse"
                }
              }
            }
          },
          "400": {
            "description": "INVALID_REQUEST",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "404": {
            "description": "COLLECTION_NOT_FOUND | ADDRESS_NOT_FOUND",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "503": {
            "description": "DATA_NOT_READY",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "500": {
            "description": "INTERNAL_ERROR",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "401": {
            "description": "UNAUTHORIZED",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/me/addresses/{address_id}": {
      "patch": {
        "tags": [
          "me"
        ],
        "summary": "Update Address",
        "description": "Đổi nhãn / mô tả / cờ mặc định. Đặt `is_default: true` thì địa chỉ mặc định cũ tự\nthôi làm mặc định (cùng một giao dịch).",
        "operationId": "update_address_api_v1_me_addresses__address_id__patch",
        "parameters": [
          {
            "name": "address_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Address Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateAddressRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserAddressResponse"
                }
              }
            }
          },
          "400": {
            "description": "INVALID_REQUEST",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "404": {
            "description": "COLLECTION_NOT_FOUND | ADDRESS_NOT_FOUND",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "503": {
            "description": "DATA_NOT_READY",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "500": {
            "description": "INTERNAL_ERROR",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "401": {
            "description": "UNAUTHORIZED",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "me"
        ],
        "summary": "Delete Address",
        "operationId": "delete_address_api_v1_me_addresses__address_id__delete",
        "parameters": [
          {
            "name": "address_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Address Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageResponse"
                }
              }
            }
          },
          "400": {
            "description": "INVALID_REQUEST",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "404": {
            "description": "COLLECTION_NOT_FOUND | ADDRESS_NOT_FOUND",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "503": {
            "description": "DATA_NOT_READY",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "500": {
            "description": "INTERNAL_ERROR",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "401": {
            "description": "UNAUTHORIZED",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AddressesData": {
        "properties": {
          "addresses": {
            "items": {
              "$ref": "#/components/schemas/UserAddressSchema"
            },
            "type": "array",
            "title": "Addresses"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          }
        },
        "type": "object",
        "required": [
          "addresses",
          "total"
        ],
        "title": "AddressesData"
      },
      "AddressesResponse": {
        "properties": {
          "data": {
            "$ref": "#/components/schemas/AddressesData"
          }
        },
        "type": "object",
        "required": [
          "data"
        ],
        "title": "AddressesResponse"
      },
...
      "CreateAddressRequest": {
        "properties": {
          "label": {
            "type": "string",
            "title": "Label",
            "description": "Nhãn ngắn: Nhà, Công ty… (1–40 ký tự)."
          },
          "address_text": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Address Text",
            "description": "Mô tả tự do, tối đa 200 ký tự."
          },
          "lat": {
            "type": "number",
            "title": "Lat",
            "description": "Vĩ độ — phải nằm trong Hà Nội."
          },
          "lng": {
            "type": "number",
            "title": "Lng",
            "description": "Kinh độ — phải nằm trong Hà Nội."
          },
          "is_default": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Is Default",
            "description": "Bỏ trống: địa chỉ ĐẦU TIÊN tự thành mặc định, các địa chỉ sau thì không."
          }
        },
        "type": "object",
        "required": [
          "label",
          "lat",
          "lng"
        ],
        "title": "CreateAddressRequest"
      },
...
      "ErrorDetail": {
        "properties": {
          "code": {
            "type": "string",
            "title": "Code"
          },
          "message": {
            "type": "string",
            "title": "Message"
          },
          "details": {
            "additionalProperties": true,
            "type": "object",
            "title": "Details",
            "default": {}
          }
        },
        "type": "object",
        "required": [
          "code",
          "message"
        ],
        "title": "ErrorDetail"
      },
      "ErrorEnvelope": {
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorDetail"
          }
        },
        "type": "object",
        "required": [
          "error"
        ],
        "title": "ErrorEnvelope"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "MessageData": {
        "properties": {
          "message": {
            "type": "string",
            "title": "Message",
            "description": "Câu thông báo hiển thị thẳng cho người dùng."
          }
        },
        "type": "object",
        "required": [
          "message"
        ],
        "title": "MessageData",
        "description": "Kết quả cho các thao tác không trả về tài nguyên nào."
      },
      "MessageResponse": {
        "properties": {
          "data": {
            "$ref": "#/components/schemas/MessageData"
          }
        },
        "type": "object",
        "required": [
          "data"
        ],
        "title": "MessageResponse"
      },
...
      "UpdateAddressRequest": {
        "properties": {
          "label": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Label"
          },
          "address_text": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Address Text"
          },
          "is_default": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Is Default"
          }
        },
        "type": "object",
        "title": "UpdateAddressRequest",
        "description": "Chỉ gửi trường muốn đổi. Gửi `address_text: null` hoặc \"\" = XOÁ mô tả."
      },
      "UserAddressResponse": {
        "properties": {
          "data": {
            "$ref": "#/components/schemas/UserAddressSchema"
          }
        },
        "type": "object",
        "required": [
          "data"
        ],
        "title": "UserAddressResponse"
      },
      "UserAddressSchema": {
        "properties": {
          "address_id": {
            "type": "string",
            "title": "Address Id"
          },
          "label": {
            "type": "string",
            "title": "Label"
          },
          "address_text": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Address Text",
            "description": "Chữ người dùng tự gõ. null = không có — KHÔNG phải geocoding."
          },
          "lat": {
            "type": "number",
            "title": "Lat"
          },
          "lng": {
            "type": "number",
            "title": "Lng"
          },
          "is_default": {
            "type": "boolean",
            "title": "Is Default"
          },
          "created_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "address_id",
          "label",
          "lat",
          "lng",
          "is_default"
        ],
        "title": "UserAddressSchema"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}